+++ Cpp/Executables/Regression/ERR_Response.cpp
@@ -8,21 +8,21 @@
 #include <FROLS_Path_Config.hpp>
 #include <Regression_Algorithm.hpp>
 
-std::string err_simulation_filename(size_t N_pop, double p_ER, size_t iter, std::string network_type) {
+std::string err_simulation_filename(uint16_t N_pop, double p_ER, uint16_t iter, std::string network_type) {
     std::stringstream ss;
     ss << FROLS::FROLS_DATA_DIR << "/ERR_Simulation_" << network_type << "_" << N_pop << "_" << p_ER << "_" << iter
        << ".csv";
     return ss.str();
 }
 
-std::string SIR_Sine_filename(size_t N_pop, double p_ER, size_t iter, std::string network_type) {
+std::string SIR_Sine_filename(uint16_t N_pop, double p_ER, uint16_t iter, std::string network_type) {
     std::stringstream ss;
     ss << FROLS::FROLS_DATA_DIR << "/SIR_Sine_Trajectory_Discrete_" << iter
        << ".csv";
     return ss.str();
 }
 
-std::string SIR_diff_filename(size_t N_pop, double p_ER, size_t iter, std::string network_type) {
+std::string SIR_diff_filename(uint16_t N_pop, double p_ER, uint16_t iter, std::string network_type) {
     std::stringstream ss;
     ss << FROLS::FROLS_DATA_DIR << "/Bernoulli_SIR_Delta_MC_" << N_pop << "_" << p_ER << "_" << iter
        << ".csv";
@@ -32,17 +32,17 @@ std::string SIR_diff_filename(size_t N_pop, double p_ER, size_t iter, std::strin
 
 
 int main(int argc, char **argv) {
-    const size_t Nx = 3;
+    const uint16_t Nx = 3;
     const std::string network_type = "SIR";
     const std::vector<std::string> colnames = {"S", "I", "R"};
-    size_t N_sims = 500; // 10000;
-    size_t N_pop = 3000;
+    uint16_t N_sims = 500; // 10000;
+    uint16_t N_pop = 3000;
     double p_ER = 1.0;
     using namespace FROLS;
     using namespace std::placeholders;
-    size_t d_max = 1;
-    size_t N_output_features = 16;
-    size_t Nu = 1;
+    uint16_t d_max = 1;
+    uint16_t N_output_features = 16;
+    uint16_t Nu = 1;
     auto Sine_fname_f = std::bind(SIR_Sine_filename, N_pop, p_ER, _1, network_type);
     auto diff_fname_f = std::bind(SIR_diff_filename, N_pop, p_ER, _1, network_type);
 
@@ -53,7 +53,7 @@ int main(int argc, char **argv) {
     model.preselect("x0", 1.0, 0, FEATURE_PRESELECTED_IGNORE);
     model.preselect("x1", 1.0, 1, FEATURE_PRESELECTED_IGNORE);
     model.preselect("x2", 1.0, 2, FEATURE_PRESELECTED_IGNORE);
-    std::vector<size_t> ignore_idx = {};//;{6, 8, 9, 13};
+    std::vector<uint16_t> ignore_idx = {};//;{6, 8, 9, 13};
     std::for_each(ignore_idx.begin(), ignore_idx.end(), [&](auto& ig_idx){
         model.ignore(ig_idx);
     });
