+++ Cpp/static/DataFrame/DataFrame.hpp
@@ -12,18 +12,18 @@ namespace FROLS {
 
     class DataFrame {
         std::map<std::string, std::shared_ptr<std::vector<double>>> data = {};
-        size_t N_rows = 100;
+        uint16_t N_rows = 100;
 
         friend class DataFrameStack;
 
     public:
         DataFrame() {}
 
-        DataFrame(size_t N_rows) : N_rows(N_rows) {}
+        DataFrame(uint16_t N_rows) : N_rows(N_rows) {}
 
         DataFrame(const std::string &filename, const std::string delimiter = ",") { read_csv(filename, delimiter); }
 
-        DataFrame(size_t N_rows, const std::vector<std::string> &col_names)
+        DataFrame(uint16_t N_rows, const std::vector<std::string> &col_names)
                 : N_rows(N_rows) {
             for (auto &col_name: col_names) {
                 data.insert({col_name, std::make_shared<std::vector<double>>(N_rows)});
@@ -38,11 +38,11 @@ namespace FROLS {
 
         const std::vector<std::vector<double>> operator[](const std::vector<std::string>& col_names) const;
 
-        void assign(const std::string &col_name, const std::vector<size_t> &col_data);
+        void assign(const std::string &col_name, const std::vector<uint16_t> &col_data);
 
         void assign(const std::string &col_name, const std::vector<double> &col_data);
 
-        template<size_t N, typename T>
+        template<uint16_t N, typename T>
         void assign(const std::string &col_name,
                     const std::array<T, N> &col_data) {
 
@@ -52,7 +52,7 @@ namespace FROLS {
             }
         }
 
-        template<size_t N, size_t N_C, typename T>
+        template<uint16_t N, uint16_t N_C, typename T>
         void assign(const std::vector<std::string> &col_names,
                     const std::array<std::array<T, N>, N_C> &data_cols) {
             for (int i = 0; i < N_C; i++) {
@@ -60,7 +60,7 @@ namespace FROLS {
             }
         }
 
-        void assign(const std::string &col_name, size_t row, double value);
+        void assign(const std::string &col_name, uint16_t row, double value);
 
         void assign(const std::string &col_name, crVec &vec);
 
@@ -70,7 +70,7 @@ namespace FROLS {
                     const std::vector<std::vector<double>> &col_data);
 
         void assign(const std::vector<std::string> &col_names,
-                    const std::vector<std::vector<size_t>> &col_data);
+                    const std::vector<std::vector<uint16_t>> &col_data);
 
 
         std::vector<std::string> get_col_names();
@@ -82,9 +82,9 @@ namespace FROLS {
                        const std::string &delimiter,
                        const double termination_tol = -std::numeric_limits<double>::infinity());
 
-        size_t get_N_rows() const { return N_rows; }
+        uint16_t get_N_rows() const { return N_rows; }
 
-        void resize(size_t N_rows);
+        void resize(uint16_t N_rows);
 
         void resize();
     };
@@ -97,7 +97,7 @@ namespace FROLS {
         std::vector<DataFrame> dataframes;
         DataFrameStack(const std::vector<std::string> &filenames) {
             dataframes.reserve(filenames.size());
-            for (size_t i = 0; i < filenames.size(); i++) {
+            for (uint16_t i = 0; i < filenames.size(); i++) {
                 dataframes.push_back(DataFrame(filenames[i]));
             }
         }
@@ -106,28 +106,28 @@ namespace FROLS {
             this->dataframes = dataframes;
         }
 
-        DataFrameStack(size_t N_frames)
+        DataFrameStack(uint16_t N_frames)
         {
             dataframes.resize(N_frames);
         }
 
-        double get_elem(size_t frame, const std::string &col_name, size_t row) {
+        double get_elem(uint16_t frame, const std::string &col_name, uint16_t row) {
             return dataframes[frame][col_name]->operator[](row);
         }
 
 
 
-        DataFrame &operator[](size_t frame) { return dataframes[frame]; }
+        DataFrame &operator[](uint16_t frame) { return dataframes[frame]; }
 
         DataFrame compute_svd();
 
-        const std::vector<double> get_col(size_t frame, const std::string &col_name);
+        const std::vector<double> get_col(uint16_t frame, const std::string &col_name);
 
-        std::vector<double> get_row(size_t frame, size_t row);
+        std::vector<double> get_row(uint16_t frame, uint16_t row);
 
-        size_t get_N_frames() const { return dataframes.size(); }
+        uint16_t get_N_frames() const { return dataframes.size(); }
 
-        std::vector<size_t> get_N_rows();
+        std::vector<uint16_t> get_N_rows();
     };
 
 } // namespace FROLS
